requirements.txt
pymupdf>=1.24.0
openai>=1.0.0
huggingface_hub>=0.20.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pytest>=7.0.0
